Add USB example code from pico SDK
Add example code to add USB HID functionality to the project.
This will serve as a working starting point for incremental
modifications to convert this application from the HID profile to the
combined MIDI/CRC profiles required for this application.

The USB operations are semi-integrated with FreeRTOS.
TinyUSB has support for FreeRTOS internally but the pico SDK seems to
block that by default and there doesn't look to be a sensible way
unblock.

For now, I have worked around this by having all functions that touch
the USB side of things operating sequentially from within the same
FreeRTOS task.
This may not be appropriate when there are more important tasks flying
around at the same time.